+++ merlin/config/results_backend.py
@@ -100,7 +100,7 @@ def get_backend_password(password_file, certs_path=None):
         # The password was given instead of the filepath.
         password = password_file.strip()
     else:
-        with open(password_filepath, "r") as f:
+        with open(password_filepath, "r") as f:  # pylint: disable=C0103
             line = f.readline().strip()
             password = quote(line, safe="")
 
@@ -151,9 +151,9 @@ def get_redis(certs_path=None, include_password=True, ssl=False):  # noqa C901
             password = CONFIG.results_backend.password
 
         if include_password:
-            spass = "%s:%s@" % (username, password)
+            spass = f"{username}:{password}@"
         else:
-            spass = "%s:%s@" % (username, "******")
+            spass = f"{username}:******@"
     except (KeyError, AttributeError):
         spass = ""
         LOG.debug(f"Results backend: redis using default password = {spass}")
@@ -183,7 +183,7 @@ def get_mysql_config(certs_path, mysql_certs):
 
     certs = {}
     for key, filename in mysql_certs.items():
-        for f in files:
+        for f in files:  # pylint: disable=C0103
             if not f == filename:
                 continue
 
@@ -215,7 +215,7 @@ def get_mysql(certs_path=None, mysql_certs=None, include_password=True):
 
     if not server:
         msg = f"Results backend: server {server} does not have a configuration"
-        raise Exception(msg)
+        raise TypeError(msg)  # TypeError since server is None and not str
 
     password = get_backend_password(password_file, certs_path=certs_path)
 
@@ -225,8 +225,9 @@ def get_mysql(certs_path=None, mysql_certs=None, include_password=True):
     mysql_config = get_mysql_config(certs_path, mysql_certs)
 
     if not mysql_config:
-        msg = f"The connection information for MySQL could not be set, cannot find:\n {mysql_certs}\ncheck the celery/certs path or set the ssl information in the app.yaml file."
-        raise Exception(msg)
+        msg = f"""The connection information for MySQL could not be set, cannot find:\n
+        {mysql_certs}\ncheck the celery/certs path or set the ssl information in the app.yaml file."""
+        raise TypeError(msg)  # TypeError since mysql_config is None when it shouldn't be
 
     mysql_config["user"] = CONFIG.results_backend.username
     if include_password:
@@ -275,16 +276,16 @@ def _resolve_backend_string(backend, certs_path, include_password):
     if "mysql" in backend:
         return get_mysql(certs_path=certs_path, include_password=include_password)
 
-    elif "sqlite" in backend:
+    if "sqlite" in backend:
         return SQLITE_CONNECTION_STRING
 
-    elif backend == "redis":
+    if backend == "redis":
         return get_redis(certs_path=certs_path, include_password=include_password)
 
-    elif backend == "rediss":
+    if backend == "rediss":
         return get_redis(certs_path=certs_path, include_password=include_password, ssl=True)
-    else:
-        return None
+
+    return None
 
 
 def get_ssl_config(celery_check=False):